Create saving and load up methods for SavingManager
