+++ AstoneTraineeWeheterApp/ViewModels/MainViewModel.swift
@@ -15,6 +15,7 @@ final class MainViewModel {
     var isPosibleToNavigate = Dynamic(false)
     var isPosibleToNavigateByLocation = Dynamic((isPosible: false, cityName: ""))
     weak var delegate: MainViewModelDelegate?
+    private let savingManager = SavingManager()
     
     // MARK: - Methods
     func searchButtonPressed(with text: String) {
@@ -23,6 +24,10 @@ final class MainViewModel {
     
     func locationButtonPressed(lon: Double, lat: Double) {
         getCurrentLocationWeather(lon: lon, lat: lat)
+    }
+    
+    func loadSavedData() -> [SearchCellViewModel] {
+        return savingManager.loadSavedLocations()
     }
 }
 
@@ -47,6 +52,8 @@ private extension MainViewModel {
                         let nightTemp = weather.main.temp_min
                         
                         self?.currentDayWeather.value = SearchCellViewModel(cityName: tempSearchText, dayTemp: Int(dayTemp.rounded(.toNearestOrAwayFromZero)), nightTepm: Int(nightTemp.rounded(.toNearestOrAwayFromZero)), wetherConditionImageID: weather.weather.first!.icon, currentTemp: Int(weather.main.temp.rounded(.toNearestOrAwayFromZero)))
+                        
+                        self?.savingManager.saveSearchedLocation( for: (self?.currentDayWeather.value)!)
                         self?.isPosibleToNavigate.value = true
                     }
                     catch {
@@ -72,6 +79,8 @@ private extension MainViewModel {
                     let nightTemp = locationWeather.main.temp_min
                     
                     self?.currentDayWeather.value = SearchCellViewModel(cityName: locationWeather.name, dayTemp: Int(dayTemp.rounded(.toNearestOrAwayFromZero)), nightTepm: Int(nightTemp.rounded(.toNearestOrAwayFromZero)), wetherConditionImageID: locationWeather.weather.first!.icon, currentTemp: Int(locationWeather.main.temp.rounded(.toNearestOrAwayFromZero)))
+                    
+                    self?.savingManager.saveSearchedLocation( for: (self?.currentDayWeather.value)!)
                     self?.isPosibleToNavigateByLocation.value = (isPosible: true, cityName: locationWeather.name)
                 }
                 catch {
